comfyui_pipeline/requirements.txt
requests>=2.31
websocket-client>=1.7
pydantic>=2.6
